refactor: removed the internal dir
