--- internal/repository/repo.go
@@ -1,5 +1,0 @@
-package repository
-
-func New() {
-
-}
