+++ manifest/manifest.go
