+++ std/command/entrypoint_test.go
