+++ std/sleep/entrypoint.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"time"
 
-	"github.com/cofunclabs/cofunc/pkg/manifest"
+	"github.com/cofunclabs/cofunc/manifest"
 )
 
 var _manifest = manifest.Manifest{
